refine branchedOptimistic strategy, cleanup tests
