+++ ark/schema/roots/basis.ts
@@ -32,7 +32,7 @@ export abstract class InternalBasis<
 		if (js.traversalKind === "Allows") js.return(this.compiledCondition)
 		else {
 			js.if(this.compiledNegation, () =>
-				js.line(`${js.ctx}.errorFromNodeContext(${this.compiledErrorContext})`)
+				js.line(`ctx.errorFromNodeContext(${this.compiledErrorContext})`)
 			)
 		}
 	}
